disco: keep transaction metadata within 40 bytes
Narrow payload_sz to ushort and group the source fields beside it so
fd_txn_p_t occupies 4992 bytes instead of 5056.  fd_txn_e_t also loses
one cache line, and the default replay transaction pool saves 4 MiB.
Keep first_seen_nanos in fd_txn_p_t for the existing telemetry path.

Assert the payload bound and compact footprint, explicitly narrow the
bounded producers, and update the scheduler footprint expectations.

Adapted from FireBAM commit 7d21e18e6ad231dbdc1d6a8c8bd48b14793a925d.

+++ src/flamenco/runtime/program/test_create_account_allow_prefund.c
@@ -169,7 +169,7 @@ build_txn( test_env_t *        env,
   memcpy( p, owner->key, 32 ); p += 32;
 
   ulong payload_sz = (ulong)(p - start);
-  env->txn_p->payload_sz = payload_sz;
+  env->txn_p->payload_sz = (ushort)payload_sz;
   FD_TEST( fd_txn_parse( env->txn_p->payload, payload_sz, TXN(env->txn_p), NULL ) > 0 );
 
   env->txn_in->txn              = env->txn_p;
